model_inference/pytorch/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.1+cu113
torchinfo==1.6.3
pytorch-ignite==0.4.7
confluent-kafka==1.8.2
avro==1.11.0
requests==2.27.1